requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
arxiv==2.1.3
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
feedparser==6.0.11
frozenlist==1.5.0
fuzzywuzzy==0.18.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langchainhub==0.1.15
langsmith==0.1.147
Levenshtein==0.26.1
marshmallow==3.23.2
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.58.1
orjson==3.10.12
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
StackAPI==0.3.1
starlette==0.41.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wikipedia==1.4.0
yarl==1.18.3
